+++ docs/CHANGELOG.rst
@@ -9,6 +9,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Added
+-----
+- Add helper function shallowPickType for getting a type with limited fields
 ==================
 8.0.1 - 2018-11-05
 ==================